@@ -1,115 +1,138 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import { setShowAbout, startConnection } from '../api/Actions';
-import { formatVersion, isCompatible, RequiredOpenSpaceVersion, RequiredSocketApiVersion } from '../api/Version';
+import {
+  formatVersion, isCompatible, RequiredOpenSpaceVersion, RequiredSocketApiVersion,
+} from '../api/Version';
 import BottomBar from '../components/BottomBar/BottomBar';
 import KeybindingPanel from '../components/BottomBar/KeybindingPanel';
+import environment from '../api/Environment'
 import Error from '../components/common/Error/Error';
 import Button from '../components/common/Input/Button/Button';
 import Overlay from '../components/common/Overlay/Overlay';
 import Stack from '../components/common/Stack/Stack';
+import LuaConsole from '../components/LuaConsole/LuaConsole';
 import NodeMetaContainer from '../components/NodeMetaPanel/NodeMetaContainer';
 import NodePopOverContainer from '../components/NodePropertiesPanel/NodePopOverContainer';
 import Sidebar from '../components/Sidebar/Sidebar';
 import '../styles/base.scss';
 import About from './About/About';
 import styles from './OnScreenGui.scss';
+import TourPopup from '../components/GettingStartedTour/TourPopup'
+import { RefsProvider } from '../components/GettingStartedTour/GettingStartedContext';
 
-class OnScreenGui extends Component {
-  constructor(props) {
-    super(props);
-    this.checkedVersion = false;
-    this.showFlightController = props.showFlightController;
+function OnScreenGui({
+  showFlightController, connectionLost, startConnection, version, showAbout, hideAbout, isInBrowser
+}) {
+  let hasCheckedVersion = false;
+  const location = useLocation();
+  const [showTutorial, setShowTutorial] = React.useState(false);
+  const [luaConsoleVisible, setLuaConsoleVisible] = React.useState(false);
+
+  React.useEffect(() => {
+    startConnection();
+    window.addEventListener("keydown", toggleConsole);
+    return () => window.removeEventListener('keydown', toggleConsole);
+  }, []);
+
+ 
+  function toggleConsole(e) {
+    if (e.code === "Backquote") {
+      setLuaConsoleVisible(current => !current);
+      }
   }
 
-  componentDidMount() {
-    this.props.startConnection();
-  }
-
-  checkVersion() {
-    if (!this.checkedVersion && this.props.version.isInitialized) {
-      const versionData = this.props.version.data;
-      if (!isCompatible(versionData.openSpaceVersion, RequiredOpenSpaceVersion)) {
-        console.warn(
-          'Possible incompatibility: \nRequired OpenSpace version: ' +
-          formatVersion(RequiredOpenSpaceVersion) +
-          '. Currently controlling OpenSpace version ' +
-          formatVersion(versionData.openSpaceVersion) + '.'
-        );
-      }
-      if (!isCompatible(versionData.socketApiVersion, RequiredSocketApiVersion)) {
-        console.warn(
-          "Possible incompatibility: \nRequired Socket API version: " +
-          formatVersion(RequiredSocketApiVersion) +
-          ". Currently operating over API version " +
-          formatVersion(versionData.socketApiVersion) + '.'
-        );
-      }
-      this.checkedVersion = true;
+  // Check the version
+  if (!hasCheckedVersion && version.isInitialized) {
+    const versionData = version.data;
+    if (!isCompatible(versionData.openSpaceVersion, RequiredOpenSpaceVersion)) {
+      console.warn(
+        `Possible incompatibility: \nRequired OpenSpace version: ${
+          formatVersion(RequiredOpenSpaceVersion)
+        }. Currently controlling OpenSpace version ${
+          formatVersion(versionData.openSpaceVersion)}.`,
+      );
     }
+    if (!isCompatible(versionData.socketApiVersion, RequiredSocketApiVersion)) {
+      console.warn(
+        `Possible incompatibility: \nRequired Socket API version: ${
+          formatVersion(RequiredSocketApiVersion)
+        }. Currently operating over API version ${
+          formatVersion(versionData.socketApiVersion)}.`,
+      );
+    }
+    hasCheckedVersion = true;
   }
 
-  reloadGui() {
+  function reloadGui() {
     location.reload();
   }
 
-  render() {
-    this.checkVersion();
-    return (
-      <div className={styles.app}>
-        { this.props.showAbout && (
+  return (
+    <div className={styles.app} style={environment.developmentMode ? { borderStyle: 'solid', borderWidth: '3px', borderColor: 'orange' } : null}>
+      {environment.developmentMode &&
+        <div className={styles.devModeTextBox}>
+          <p>Dev Gui</p>
+        </div>
+      }
+      <RefsProvider>
+        { showAbout && (
+
           <Overlay>
             <Stack style={{ maxWidth: '500px' }}>
-              <Button style={{ alignSelf: 'flex-end', color: 'white' }} onClick={this.props.hideAbout}>
+              <Button style={{ alignSelf: 'flex-end', color: 'white' }} onClick={hideAbout}>
                 Close
               </Button>
               <About />
             </Stack>
           </Overlay>
         )}
-        { this.props.connectionLost && (
+        { connectionLost && (
           <Overlay>
             <Error>
               <h2>Houston, we've had a...</h2>
               <p>...disconnection between the user interface and OpenSpace.</p>
               <p>Trying to reconnect automatically, but you may want to...</p>
-              <Button className={Error.styles.errorButton} onClick={this.reloadGui}>Reload the user interface</Button>
+              <Button className={Error.styles.errorButton} onClick={reloadGui}>Reload the user interface</Button>
             </Error>
           </Overlay>
         )}
         <section className={styles.Grid__Left}>
-          <Sidebar />
+          <Sidebar showTutorial={ setShowTutorial } />
         </section>
         <section className={styles.Grid__Right}>
+          {isInBrowser && luaConsoleVisible && <LuaConsole />}
           <NodePopOverContainer />
           <NodeMetaContainer />
-          <BottomBar showFlightController={this.props.showFlightController}/>
+          <BottomBar showFlightController={showFlightController} />
           <KeybindingPanel />
+          <TourPopup isVisible={showTutorial} setVisibility = { (show) => setShowTutorial(show)}/>
         </section>
-      </div>
-    );
-  }
+      </RefsProvider>
+    </div>
+  );  
+
 }
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state) => ({
   connectionLost: state.connection.connectionLost,
   version: state.version,
-  showAbout: state.local.showAbout
+  showAbout: state.local.showAbout,
 });
 
-const mapDispatchToProps = dispatch => ({
+const mapDispatchToProps = (dispatch) => ({
   startConnection: () => {
     dispatch(startConnection());
   },
   hideAbout: () => {
-    dispatch(setShowAbout(false))
-  }
+    dispatch(setShowAbout(false));
+  },
 });
 
-OnScreenGui = withRouter(connect(
+OnScreenGui = connect(
   mapStateToProps,
   mapDispatchToProps,
-)(OnScreenGui));
+)(OnScreenGui);
 
 export default OnScreenGui;
