@@ -9,6 +9,8 @@ import Popover from '../common/Popover/Popover';
 import Row from '../common/Row/Row';
 import styles from './ActionsPanel.scss';
 import Picker from './Picker';
+import { FilterList, FilterListData, FilterListFavorites } from '../common/FilterList/FilterList';
+import { ObjectWordBeginningSubstring } from '../../utils/StringMatchers';
 
 class ActionsPanel extends Component {
   constructor(props) {
@@ -20,25 +22,29 @@ class ActionsPanel extends Component {
     this.getActionContent = this.getActionContent.bind(this);
     this.getChildrenContent = this.getChildrenContent.bind(this);
     this.getBackButton = this.getBackButton.bind(this);
+    this.getAllActions = this.getAllActions.bind(this);
+    this.matcher = this.matcher.bind(this);
   }
 
   togglePopover() {
-    this.props.setPopoverVisibility(!this.props.popoverVisible)
+    this.props.setPopoverVisibility(!this.props.popoverVisible);
   }
 
   addNavPath(e) {
-    var navString = this.props.navigationPath;
-    if (this.props.navigationPath == '/') {
+    const { navigationPath } = this.props;
+    let navString = navigationPath;
+    if (navigationPath == '/') {
       navString += e.currentTarget.getAttribute('foldername');
     } else {
-      navString +=  "/" + e.currentTarget.getAttribute('foldername');
+      navString += `/${e.currentTarget.getAttribute('foldername')}`;
     }
     this.props.actionPath(navString);
   }
 
   goBack(e) {
-    var navString = this.props.navigationPath;
-    navString = navString.substring(0,navString.lastIndexOf('/'));
+    const { navigationPath } = this.props;
+    let navString = navigationPath;
+    navString = navString.substring(0, navString.lastIndexOf('/'));
     if (navString.length == 0) {
       navString = '/';
     }
@@ -46,104 +52,122 @@ class ActionsPanel extends Component {
   }
 
   sendAction(e) {
-    var actionId = e.currentTarget.getAttribute('actionid');
-    this.props.luaApi.action.triggerAction(actionId);
+    const { luaApi } = this.props;
+    const actionId = e.currentTarget.getAttribute('actionid');
+    luaApi.action.triggerAction(actionId);
   }
 
   getActionContent(level) {
-    return level.actions.map(action =>
-        <Button
-          block
-          smalltext
-          onClick={this.sendAction}
-          key={action.identifier}
-          className={styles.actionButton}
-          actionid={action.identifier}
-        >
-          <p><MaterialIcon className={styles.buttonIcon} icon="launch" /></p>
-          {action.name}
-          <InfoBox inpanel panelscroll='actionscroller' text={action.documentation} />
-        </Button>
-      );
+    return level.actions.map((action) => (
+      <Button
+        block
+        smalltext
+        onClick={this.sendAction}
+        key={action.identifier}
+        className={styles.actionButton}
+        actionid={action.identifier}
+      >
+        <p><MaterialIcon className={styles.buttonIcon} icon="launch" /></p>
+        {action.name}
+        {action.documentation && <InfoBox text={action.documentation} />}
+      </Button>
+    ));
   }
 
   getChildrenContent(level) {
-    return Object.keys(level.children).map(key =>
-        <Button
-          block
-          smalltext
-          onClick={this.addNavPath}
-          key={key}
-          foldername={key}
-          className={`${styles.actionButton} ${styles.folderButton}`}
-        >
-          <p><MaterialIcon className={styles.buttonIcon} icon="folder" /></p>
-          {key}
-        </Button>
-      );
+    return Object.keys(level.children).map((key) => (
+      <Button
+        block
+        smalltext
+        onClick={this.addNavPath}
+        key={key}
+        foldername={key}
+        className={`${styles.actionButton} ${styles.folderButton}`}
+      >
+        <p><MaterialIcon className={styles.buttonIcon} icon="folder" /></p>
+        {key}
+      </Button>
+    ));
+  }
+
+  getAllActions() {
+    const { allActions } = this.props;
+    return allActions.map((action) => (
+      <Button
+        block
+        smalltext
+        onClick={this.sendAction}
+        key={`${action.identifier}Filtered`}
+        className={styles.actionButton}
+        actionid={action.identifier}
+        name={action.name}
+      >
+        <p><MaterialIcon className={styles.buttonIcon} icon="launch" /></p>
+        {action.name}
+      </Button>
+    ));
   }
 
   getBackButton() {
-    if (this.props.navigationPath != '/') {
-        return <Button block smalltext className={styles.backButton} onClick={this.goBack} key='backbtn' >&lt;- Back</Button>;
+    const { navigationPath } = this.props;
+    if (navigationPath != '/') {
+      return <Button block className={styles.backButton} onClick={this.goBack} key="backbtn">
+        <MaterialIcon className={styles.buttonIcon} icon="arrow_back" />
+      </Button>;
     }
   }
 
-  get windowContent() {
-   var level = this.props.actionLevel;
+  matcher(test, search) {
+    return ObjectWordBeginningSubstring(test.children, search);
+  }
 
-   if (level == undefined) {
-     return <div>Loading</div>
-   }
-   var actionsContent = this.getActionContent(level);
-   var childrenContent = this.getChildrenContent(level);
-   var backButton = this.getBackButton();
-   return (
-    <div id='actionscroller' className={styles.windowContainer}>
-      <hr className={Popover.styles.delimiter} />
-      <Row>
-        <div>{this.props.navigationPath} </div>
-      </Row>
-      <hr className={Popover.styles.delimiter} />
-      <div className={styles.Grid}>
+  get windowContent() {
+    const { displayedNavigationPath, level } = this.props;
+
+    if (level == undefined) {
+      return <div>Loading</div>;
+    }
+    const actionsContent = this.getActionContent(level);
+    const childrenContent = this.getChildrenContent(level);
+    const backButton = this.getBackButton();
+    
+    return (
+      <div id="actionscroller" className={`${styles.windowContainer}`}>
+        <hr className={Popover.styles.delimiter} />
+          <Row>
             {backButton}
+            <div className={styles.navPathTitle}>
+              {`${displayedNavigationPath}`}
+            </div>
+          </Row>
+        <hr className={Popover.styles.delimiter} />
+        <FilterList matcher={this.matcher} height={'80%'}>
+          <FilterListFavorites className={styles.Grid}>
             {actionsContent}
             {childrenContent}
-       </div>
+          </FilterListFavorites>
+          <FilterListData className={styles.Grid}>
+            {this.getAllActions()}
+          </FilterListData>
+        </FilterList>
       </div>
     );
   }
 
   get popover() {
-    var actionsContent;
-    var childrenContent;
-    var backButton;
-    var keybindsContent;
+    const { actionLevel, displayedNavigationPath } = this.props;
+    let actionsContent;
+    let childrenContent;
+    let backButton;
 
-    if (this.props.actionLevel.length == 0) {
+    if (actionLevel.length == 0) {
       actionsContent = <div>No Actions</div>;
       childrenContent = <div>No Children</div>;
     } else {
-      var level = this.props.actionLevel;
+      const level = actionLevel;
       actionsContent = this.getActionContent(level);
       childrenContent = this.getChildrenContent(level);
       backButton = this.getBackButton();
-    }
-
-    var navPathString = this.props.navigationPath;
-
-    if (navPathString == '/') {
-      keybindsContent = <Button
-          block
-          smalltext
-          onClick={this.props.toggleKeybinds}
-          key="showKeybinds"
-          className={styles.actionButton}
-        >
-          <p><MaterialIcon className={styles.buttonIcon} icon="launch" /></p>
-          Show Keybindings
-          <InfoBox inpanel panelscroll='actionscroller' text="Shows the keybinding vieiwer" />
-        </Button>;
     }
 
     return (
@@ -152,76 +176,94 @@ class ActionsPanel extends Component {
         title="Actions"
         closeCallback={this.togglePopover}
         detachable
-        attached={true}
-      >        
-        <div id='actionscroller' className={`${Popover.styles.content} ${styles.scroller}` }>
+        attached
+      >
+        <div id="actionscroller" className={`${Popover.styles.content}`}>
           <hr className={Popover.styles.delimiter} />
-          <Row>
-            <div className={Popover.styles.title}>{navPathString} </div>
+          <Row className={styles.navPathRow}>
+            {backButton}
+            <div className={styles.navPathTitle}>
+              {`${displayedNavigationPath}`}
+            </div>
           </Row>
           <hr className={Popover.styles.delimiter} />
-          <div className={styles.Grid}>
-            {backButton}
-            {actionsContent}
-            {childrenContent}
-            {keybindsContent}
-          </div>
+          <FilterList matcher={this.matcher} height={backButton ? '280px' : '320px'}>
+            <FilterListFavorites className={styles.Grid}>
+              {childrenContent}
+              {actionsContent}
+            </FilterListFavorites>
+            <FilterListData className={styles.Grid}>
+              {this.getAllActions()}
+            </FilterListData>
+          </FilterList>
         </div>
       </Popover>
     );
   }
 
   render() {
-    const { popoverVisible, actionLevel, navigationPath, singlewindow } = this.props;
+    const { popoverVisible, singlewindow } = this.props;
     if (singlewindow) {
-      return (
-        <div>
-        { this.windowContent }
-        </div>
-      );
-    } else {
-      return (
-        <div className={Picker.Wrapper}>
-          <Picker 
-            className={`${popoverVisible && Picker.Active}`} 
-            onClick={this.togglePopover}
-          >
-            <div>
-              <MaterialIcon className={styles.bottomBarIcon} icon="dashboard" />
-            </div>
-          </Picker>
-          { popoverVisible && this.popover }
-        </div>
-      );
-
+      return (this.windowContent);
     }
+    return (
+      <div className={Picker.Wrapper}>
+        <Picker
+          refKey="Actions"
+          className={`${popoverVisible && Picker.Active}`}
+          onClick={this.togglePopover}
+        >
+          <div>
+            <MaterialIcon className={styles.bottomBarIcon} icon="dashboard" />
+          </div>
+        </Picker>
+        { popoverVisible && this.popover }
+      </div>
+    );
   }
 }
 
-
-const mapSubStateToProps = ({popoverVisible, luaApi, actions}) => {
-  var actionsMapped = {"/": {actions:[], children:{}}};
+const mapSubStateToProps = ({ popoverVisible, luaApi, actions }) => {
+  const actionsMapped = { '/': { actions: [], children: {} } };
   if (!actions.isInitialized) {
     return {
       actions: actionsMapped,
-      popoverVisible: popoverVisible,
-      luaApi: luaApi,
-      navigationPath: '/'
-    }
+      popoverVisible,
+      luaApi,
+      navigationPath: '/',
+      displayedNavigationPath: '/'
+    };
   }
 
-  for (var i = 0; i < actions.data.shortcuts.length; i++) {
-    var action = actions.data.shortcuts[i];
+  for (let i = 0; i < actions.data.shortcuts.length; i++) {
+    const action = actions.data.shortcuts[i];
     if (action.key) {
       continue;
     }
-    var splits = action.guiPath.split('/');
-    splits.shift();
-    var parent = actionsMapped['/'];
+
+    // If there is no backslash at beginning of GUI path, add that manually
+    // (there should always be though)
+    if (action.guiPath.length > 0 && action.guiPath[0] !== '/') {
+      action.guiPath = '/' + action.guiPath;
+    }
+
+    let splits = action.guiPath.split('/');
+    // Remove all empty strings: which is what we get before initial slash and
+    // if the path is just a slash
+    splits = splits.filter((s) => s !== "");
+
+    let parent = actionsMapped['/'];
+
+    // Add to top level actions (no gui path)
+    if (splits.length == 0) {
+      parent.actions.push(action);
+    }
+
+    // Add actions of other levels
     while (splits.length > 0) {
       var index = splits.shift();
       if (parent.children[index] == undefined) {
-        parent.children[index] = {actions:[], children:{}};
+        parent.children[index] = { actions: [], children: {} };
       }
       if (splits.length == 0) {
         parent.children[index].actions.push(action);
@@ -231,8 +273,8 @@ const mapSubStateToProps = ({popoverVisible, luaApi, actions}) => {
     }
   }
 
-  var navPath = actions.navigationPath;
-  var actionsForPath = actionsMapped['/'];
+  const navPath = actions.navigationPath;
+  let actionsForPath = actionsMapped['/'];
   if (navPath.length > 1) {
     var splits = navPath.split('/');
     splits.shift();
@@ -242,12 +284,49 @@ const mapSubStateToProps = ({popoverVisible, luaApi, actions}) => {
     }
   }
 
+  const allActions = actions.data.shortcuts.filter(action => {
+    if (navPath.length === 1) {
+      return true;
+    }
+    else {
+      var navPathGui = navPath.split('/');
+      var actionPathGui = action.guiPath?.split('/');
+      for (let i = 0; i < navPathGui.length; i++) {
+        if (actionPathGui?.[i] !== navPathGui[i]) {
+          return false;
+        }
+      }
+      return true;
+    }
+  });
+
+  // Truncate navigation path if too long
+  const NAVPATH_LENGTH_LIMIT = 60;
+  const shouldTruncate = actions.navigationPath.length > NAVPATH_LENGTH_LIMIT;
+
+  let truncatedPath = actions.navigationPath;
+  if (shouldTruncate) {
+    let originalPath = navPath;
+    if (originalPath[0] === '/') {
+      originalPath = originalPath.substring(1);
+    }
+    let pieces = originalPath.split('/');
+    if (pieces.length > 1) {
+      // TODO: maybe keep more pieces of the path, if possible?
+      truncatedPath = `/${pieces[0]}/.../${pieces[pieces.length - 1]}`;
+    } else {
+      truncatedPath = navPath.substring(0, NAVPATH_LENGTH_LIMIT);
+    }
+  }
+
   return {
     actionLevel: actionsForPath,
-    popoverVisible: popoverVisible,
-    luaApi: luaApi,
-    navigationPath: actions.navigationPath
-  }
+    popoverVisible,
+    luaApi,
+    navigationPath: actions.navigationPath,
+    displayedNavigationPath: truncatedPath,
+    allActions: allActions,
+  };
 };
 
 const mapStateToSubState = (state) => ({
@@ -256,11 +335,11 @@ const mapStateToSubState = (state) => ({
   actions: state.shortcuts,
 });
 
-const mapDispatchToProps = dispatch => ({
-  setPopoverVisibility: visible => {
+const mapDispatchToProps = (dispatch) => ({
+  setPopoverVisibility: (visible) => {
     dispatch(setPopoverVisibility({
       popover: 'actions',
-      visible
+      visible,
     }));
   },
   trigger: (action) => {
@@ -269,17 +348,11 @@ const mapDispatchToProps = dispatch => ({
   actionPath: (action) => {
     dispatch(setActionsPath(action));
   },
-  toggleKeybinds: (action) => {
-    dispatch(setPopoverVisibility({
-      popover: 'keybinds',
-      visible: true
-    }));
-  }
-})
+});
 
 ActionsPanel = connect(
   subStateToProps(mapSubStateToProps, mapStateToSubState),
-  mapDispatchToProps
+  mapDispatchToProps,
 )(ActionsPanel);
 
 export default ActionsPanel;
