@@ -1,13 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styles from './SkyBrowserTooltip.scss';
-import { excludeKeys } from '../../../utils/helpers';
 
-const SkyBrowserTooltip = (props) => {
-  const { children, placement, fixed } = props;
+function SkyBrowserTooltip({ children, fixed, placement, ...props }) {
   return (
     <div
-      {...excludeKeys(props, 'placement fixed')}
+      {...props}
       className={`${styles.tooltip} ${styles[placement]} ${fixed && styles.fixed}`}
     >
       <div className={styles.content}>
@@ -19,8 +17,8 @@ const SkyBrowserTooltip = (props) => {
 
 SkyBrowserTooltip.propTypes = {
   children: PropTypes.node.isRequired,
-  placement: PropTypes.oneOf('bottom-left bottom-right'.split(' ')),
   fixed: PropTypes.bool,
+  placement: PropTypes.oneOf('bottom-left bottom-right left'.split(' ')),
 };
 
 SkyBrowserTooltip.defaultProps = {
