@@ -1,100 +1,108 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import shallowEqualArrays from 'shallow-equal/arrays';
-import shallowEqualObjects from 'shallow-equal/objects';
+import React from 'react';
+import { useSelector, useDispatch, shallowEqual } from 'react-redux';
+import { lowPrecisionEqual } from '../../../utils/customHooks';
 import Picker from '../Picker';
+import PropTypes from 'prop-types';
 import FloatingWindow from './WindowThreeStates/FloatingWindow'
 import styles from './WorldWideTelescope.scss'
 import { SkyBrowser_ShowTitleInBrowserKey, SkyBrowser_InverseZoomDirectionKey } from '../../../api/keys';
 import { getBoolPropertyValue } from '../../../utils/propertyTreeHelpers';
+import {
+  subscribeToProperty,
+  unsubscribeToProperty
+} from '../../../api/Actions';
 
-class WorldWideTelescope extends Component {
-  constructor(props) {
-    super(props);
-    this.iframe = React.createRef();
-    this.state = {
-      isDragging: false,
-      startDragPosition: [0,0],
-      topBarHeight: 25,
-    };
-    this.sendMessageToWwt = this.sendMessageToWwt.bind(this);
-    this.setAim = this.setAim.bind(this);
-    this.mouseDown = this.mouseDown.bind(this);
-    this.mouseUp = this.mouseUp.bind(this);
-    this.handleDrag = this.handleDrag.bind(this);
-    this.handleCallbackMessage = this.handleCallbackMessage.bind(this);
-    this.scroll = this.scroll.bind(this);
-    this.changeSize = this.changeSize.bind(this);
-    this.setBorderColor = this.setBorderColor.bind(this);
-    this.color = [255, 255, 255];
-    this.ratio = 1;
-  }
+function WorldWideTelescope({ 
+  imageCollectionIsLoaded,
+  position,
+  setImageCollectionIsLoaded,
+  setMessageFunction,
+  setPosition,
+  setSize,
+  size,
+  togglePopover
+}) {
+  // State
+  const [isDragging, setIsDragging] = React.useState(false);
+  const [startDragPosition, setStartDragPosition] = React.useState([0, 0]);
+  const [wwtHasLoaded, setWwtHasLoaded] = React.useState(false);
+  const [setupWwtFunc, setSetupWwtFunc] = React.useState(null);
+  const TopBarHeight = 25;
+  // Refs
+  const iframe = React.useRef(null);
 
-  componentDidMount() {
-    this.props.setMessageFunction(this.sendMessageToWwt);
-    window.addEventListener("message", this.handleCallbackMessage);
-    this.props.setImageCollectionIsLoaded(false);
-  }
+  // Selectors & dispatch - access Redux store
+  // Get each value separately to reduce unnecessary renders
+  const fov = useSelector((state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].fov, lowPrecisionEqual);
+  const ra = useSelector((state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].ra, lowPrecisionEqual);
+  const dec = useSelector((state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].dec, lowPrecisionEqual);
+  const roll = useSelector((state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].roll, lowPrecisionEqual);
+  const borderRadius = useSelector(
+    (state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].borderRadius
+  , lowPrecisionEqual);
+  const browserName = useSelector((state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].name);
+  const browserId = useSelector((state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].id);
+  const browserColor = useSelector((state) => state.skybrowser.browsers[state.skybrowser.selectedBrowserId].color, shallowEqual);
+  const url = useSelector((state) => state.skybrowser.url);
+  const skybrowserApi = useSelector((state) => state.luaApi.skybrowser);
+  const showTitle = useSelector((state) => getBoolPropertyValue(state, SkyBrowser_ShowTitleInBrowserKey));
+  const inverseZoom = useSelector((state) => getBoolPropertyValue(state, SkyBrowser_InverseZoomDirectionKey));
+  const BorderWidth = 10;
 
-  componentWillUnmount() {
-    window.removeEventListener("message", this.handleCallbackMessage);
-  }
+  const dispatch = useDispatch();
 
-  componentDidUpdate(prevProps) {
-    const { browserAimInfo, browserColor } = this.props;
-    if (prevProps.browserColor !== browserColor) {
-      this.setBorderColor(browserColor);
-      this.color = browserColor;
+  // Effects
+  React.useEffect(() => {
+    setMessageFunction(sendMessageToWwt);
+    window.addEventListener("message", handleCallbackMessage);
+    setImageCollectionIsLoaded(false);
+    return () => window.removeEventListener("message", handleCallbackMessage);
+  }, []);
+
+  React.useEffect(() => {
+    // Send aim messages to WorldWide Telescope to prompt it to reply with a message
+    setSetupWwtFunc(setInterval(setAim, 250));
+  }, []);
+
+  React.useEffect(() => {
+    if (wwtHasLoaded && url !== '' && !imageCollectionIsLoaded) {
+      initialize();
     }
+  }, [wwtHasLoaded, url]);
+ 
+  React.useEffect(() => {
+    setAim();
+  }, [fov, roll, ra, dec]);
 
-    if (!shallowEqualObjects(prevProps.browserAimInfo, browserAimInfo)) {
-      this.setAim(browserAimInfo);
-    }
-  }
+  React.useEffect(() => {
+    setBorderRadius(borderRadius);
+  }, [borderRadius]);
 
-  shouldComponentUpdate(nextProps, nextState) {
-    const { browserAimInfo, browserColor, browserId } = this.props;
+  React.useEffect(() => {
+    setBorderColor(browserColor);
+  }, [browserColor]);
 
-    return (
-      !shallowEqualObjects(nextProps.browserAimInfo, browserAimInfo) ||
-      nextProps.browserId !== browserId ||
-      !shallowEqualArrays(nextProps.browserColor,browserColor)
-    );
-  }
+  React.useEffect(() => {
+    dispatch(subscribeToProperty(SkyBrowser_ShowTitleInBrowserKey));
+    dispatch(subscribeToProperty(SkyBrowser_InverseZoomDirectionKey));
+    return () => {
+      dispatch(unsubscribeToProperty(SkyBrowser_ShowTitleInBrowserKey));
+      dispatch(unsubscribeToProperty(SkyBrowser_InverseZoomDirectionKey));
+    }; 
+  }, []);
 
-  handleCallbackMessage(event) {
-    const { browserId, url } = this.props;
-    if (event.data == "wwt_has_loaded") {
-      this.sendMessageToWwt({
-        event: "modify_settings",
-        settings: [["hideAllChrome", true]],
-        target: "app"
-      });
-      this.sendMessageToWwt({
-        event: "load_image_collection",
-        url: url,
-        loadChildFolders: true
-      });
-      this.setBorderColor(this.props.browserColor);
-    }
-    if (event.data == "load_image_collection_completed") {
-      this.props.setImageCollectionIsLoaded(true);
-      this.props.addAllSelectedImages(browserId, false);
-    }
-  }
-
-  setBorderColor(color) {
-    this.color = color;
-    this.sendMessageToWwt({
-      event: "set_background_color",
-      data: color
+  // When WorldWide Telescope has replied with a message, stop sending it unnecessary messages
+  function stopSetupWwtFunc() {
+    setSetupWwtFunc(interval => {
+        clearInterval(interval);
+        return null;
     });
   }
 
-  sendMessageToWwt(message) {
+  function sendMessageToWwt(message) {
     try {
-      let frame = this.iframe.current.contentWindow;
-      if (frame) {
+      let frame = iframe.current.contentWindow;
+      if (frame && message) {
         frame.postMessage(message, "*");
       }
     } catch (error) {
@@ -102,135 +110,161 @@ class WorldWideTelescope extends Component {
     }
   }
 
-  setAim(aimInfo) {
-    this.sendMessageToWwt({
+  function initialize() {
+    sendMessageToWwt({
+        event: "modify_settings",
+        settings: [["hideAllChrome", true]],
+        target: "app"
+      });
+      sendMessageToWwt({
+        event: "load_image_collection",
+        url: url,
+        loadChildFolders: true
+      });
+      setBorderColor(browserColor);
+      setBorderRadius(borderRadius);
+  }
+
+  function handleCallbackMessage(event) {
+    if (event.data == "wwt_has_loaded") {
+      setWwtHasLoaded(true);
+      stopSetupWwtFunc();
+    }
+    if (event.data == "load_image_collection_completed") {
+      setImageCollectionIsLoaded(true);
+    }
+  }
+
+  function setBorderRadius(radius) {
+    sendMessageToWwt({
+      event: "set_border_radius",
+      data: radius
+    });
+  }
+
+  function setBorderColor(color) {
+    sendMessageToWwt({
+      event: "set_background_color",
+      data: color
+    });
+  }
+
+  function setAim() {
+    sendMessageToWwt({
       "event": "center_on_coordinates",
-      "ra": aimInfo.ra,
-      "dec": aimInfo.dec,
-      "fov": aimInfo.fov,
-      "roll": aimInfo.roll,
+      "ra": ra,
+      "dec": dec,
+      "fov": fov,
+      "roll": roll,
       "instant": true
     });
   }
 
-  handleDrag(mouse) {
-    if (this.state.isDragging) {
+  function handleDrag(mouse) {
+    if (isDragging) {
+      // Calculate pixel translation
       const end = [mouse.clientX, mouse.clientY];
-      this.props.skybrowserApi.finetuneTargetPosition(
-        this.props.browserId,
-        this.state.startDragPosition,
-        end
+      const width = size.width - BorderWidth;
+      const height = size.height - BorderWidth;
+      const translation = [end[0] - startDragPosition[0], end[1] - startDragPosition[1]];
+      // Calculate [ra, dec] translation without roll
+      const percentageTranslation = [translation[0] / width, translation[1] / height];  
+      // Call lua function
+      skybrowserApi.finetuneTargetPosition(
+        browserId,
+        percentageTranslation
       );
     }
   }
 
-  mouseDown(mouse) {
-    this.props.skybrowserApi.startFinetuningTarget(this.props.browserId);
+  function mouseDown(mouse) {
+    skybrowserApi.startFinetuningTarget(browserId);
     const position = [mouse.clientX, mouse.clientY];
-    this.setState({
-      isDragging: true,
-      startDragPosition: position
-    });
-    this.props.skybrowserApi.stopAnimations(this.props.browserId);
+    setIsDragging(true);
+    setStartDragPosition(position);
+    skybrowserApi.stopAnimations(browserId);
   }
 
-  mouseUp(mouse) {
-    this.setState({isDragging: false});
+  function mouseUp(mouse) {
+    setIsDragging(false);
   }
 
-  scroll(e) {
+  function scroll(e) {
     let scroll = e.deltaY;
-    if (this.props.inverseZoom) {
+    if (inverseZoom) {
       scroll *= -1;
     }
-    this.props.skybrowserApi.scrollOverBrowser(this.props.browserId, scroll);
-    this.props.skybrowserApi.stopAnimations(this.props.browserId);
+    skybrowserApi.scrollOverBrowser(browserId, scroll);
+    skybrowserApi.stopAnimations(browserId);
   }
 
-  changeSize(widthWwt, heightWwt) {
-    const { topBarHeight } = this.state;
+  function changeSize(widthWwt, heightWwt) {
     const { innerWidth: windowWidth, innerHeight: windowHeight } = window;
-    const ratio = widthWwt / (heightWwt - topBarHeight);
-    const scale = (heightWwt - topBarHeight) / windowHeight;
+    const ratio = widthWwt / (heightWwt - TopBarHeight);
+    const scale = (heightWwt - TopBarHeight) / windowHeight;
     const newWidth = 2 * scale * ratio;
     const newHeight = 2 * scale;
-    const id = this.props.browserId;
-    this.props.setSize({ width: widthWwt, height: heightWwt });
-    this.props.skybrowserApi.setBrowserRatio(id, newWidth / newHeight);
+    const id = browserId;
+    setSize({ width: widthWwt, height: heightWwt });
+    skybrowserApi.setBrowserRatio(id, newWidth / newHeight);
   }
 
-  render() {
-    const { browserName, showTitle, size, position, setPosition } = this.props;
+  const topBar =
+    <header className={`header ${styles.topMenu}`}>
+      <div className={styles.title}>{showTitle && browserName}</div>
+    </header>;
 
-    const topBar =
-      <header className={`header ${styles.topMenu}`}>
-        <div className={styles.title}>{showTitle && browserName}</div>
-      </header>;
+  // Covering div to handle interaction
+  const interactionDiv = <div
+    className={styles.container}
+    onMouseMove={handleDrag}
+    onMouseDown={mouseDown}
+    onMouseUp={mouseUp}
+    onMouseLeave={mouseUp}
+    onWheel = {(e) => scroll(e)}
+  />
 
-    // Covering div to handle interaction
-    const interactionDiv = <div
-      className={styles.container}
-      onMouseMove={this.handleDrag}
-      onMouseDown={this.mouseDown}
-      onMouseUp={this.mouseUp}
-      onMouseLeave={this.mouseUp}
-      onWheel = {(e) => this.scroll(e)}
-    />
-
-    return (
-      <FloatingWindow
-        className={`${Picker.Popover}`}
-        title={browserName}
-        closeCallback={this.togglePopover}
-        defaultSize={{ height: `425px`, width: `400px` }}
-        size={{ height: `${size.height}px`, width: `${size.width}px` }}
-        position={position}
-        handleStop={setPosition}
-        setNewHeight={this.changeSize}
-      >
-        {topBar}
-        <div className={styles.content}>
-          {interactionDiv}
-          <iframe
-            id="webpage"
-            name = "wwt"
-            ref={this.iframe}
-            src="http://wwt.openspaceproject.com/1/gui/"
-            allow="accelerometer; clipboard-write; gyroscope"
-            allowFullScreen
-            frameBorder="0"
-            align="middle"
-            className={styles.wwt}
-          >
-            <p>ERROR: cannot display AAS WorldWide Telescope research app!</p>
-          </iframe>
-        </div>
-      </FloatingWindow>
-    );
-  }
+  return (
+    <FloatingWindow
+      className={`${Picker.Popover}`}
+      title={browserName}
+      closeCallback={togglePopover}
+      defaultSize={{ height: `425px`, width: `400px` }}
+      size={{ height: `${size.height}px`, width: `${size.width}px` }}
+      position={position}
+      handleStop={setPosition}
+      setNewHeight={changeSize}
+    >
+      {topBar}
+      <div className={styles.content}>
+        {interactionDiv}
+        <iframe
+          id="webpage"
+          name = "wwt"
+          ref={el => iframe.current = el}
+          src="http://wwt.openspaceproject.com/1/gui/"
+          allow="accelerometer; clipboard-write; gyroscope"
+          allowFullScreen
+          frameBorder="0"
+          align="middle"
+          className={styles.wwt}
+        >
+          <p>ERROR: cannot display AAS WorldWide Telescope research app!</p>
+        </iframe>
+      </div>
+    </FloatingWindow>
+  );
 }
 
-const mapStateToProps = (state) => {
-  return {
-    showTitle: getBoolPropertyValue(state, SkyBrowser_ShowTitleInBrowserKey),
-    inverseZoom: getBoolPropertyValue(state, SkyBrowser_InverseZoomDirectionKey),
-  }
+WorldWideTelescope.propTypes = {
+  imageCollectionIsLoaded: PropTypes.bool,
+  position: PropTypes.object,
+  setImageCollectionIsLoaded: PropTypes.func,
+  setMessageFunction: PropTypes.func,
+  setPosition: PropTypes.func,
+  setSize: PropTypes.func,
+  size: PropTypes.object,
+  togglePopover: PropTypes.func
 };
-
-const mapDispatchToProps = dispatch => ({
-  startListeningToProperties: () => {
-    dispatch(subscribeToProperty(SkyBrowser_ShowTitleInBrowserKey));
-    dispatch(subscribeToProperty(SkyBrowser_InverseZoomDirectionKey));
-  },
-  stopListeningToProperties: () => {
-    dispatch(unsubscribeToProperty(SkyBrowser_ShowTitleInBrowserKey));
-    dispatch(unsubscribeToProperty(SkyBrowser_InverseZoomDirectionKey));
-  },
-})
-
-WorldWideTelescope = connect(
-  mapStateToProps,
-  mapDispatchToProps)
-(WorldWideTelescope);
 
 export default WorldWideTelescope;
