@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import MaterialIcon from '../MaterialIcon/MaterialIcon';
 import Tooltip from '../Tooltip/Tooltip';
+import { excludeKeys } from '../../../utils/helpers';
 
 class InfoBox extends Component {
   constructor(props) {
@@ -22,14 +23,9 @@ class InfoBox extends Component {
 
   get position() {
     if (!this.wrapper) return { top: '0px', left: '0px' };
-    if (this.props.inpanel) {
-      var scrollParent = document.getElementById(this.props.panelscroll);
-      var buttonScroll = scrollParent.scrollTop;
-      var rect = this.wrapper.getBoundingClientRect();
-      return { top: `${this.wrapper.offsetTop - buttonScroll}px`, left: `${this.wrapper.offsetLeft + rect.width}px` };
-    } else {
+    else {
       const { top, right } = this.wrapper.getBoundingClientRect();
-      return { position: "fixed", top: `${top}px`, left: `${right}px` };
+      return { top: `${top}px`, left: `${right}px` };
     }
   }
 
@@ -42,15 +38,18 @@ class InfoBox extends Component {
   }
 
   render() {
-    const { icon, text, inpanel } = this.props;
+    const { icon, text } = this.props;
     const { showPopup } = this.state;
+    const passOnProps = excludeKeys(this.props, 'icon text'); 
     return (
       <span
         ref={this.setRef('wrapper')}
+        {...passOnProps}
       >
         <MaterialIcon icon={icon}
          onMouseEnter={this.showPopup}
-         onMouseLeave={this.hidePopup} />
+         onMouseLeave={this.hidePopup} 
+        />
         { showPopup && (
           <Tooltip fixed placement="right" style={this.position}>
             { text }
