@@ -18,10 +18,10 @@ class Checkbox extends Component {
   }
 
   render() {
-    const { checked, label, wide, left, className } = this.props;
+    const { checked, label, wide, left, className, setChecked, ...rest } = this.props;
 
     return (
-      <div onClick={this.onClick} className={`${styles.wrapper} ${className} ${wide ? styles.wide : ''} ${left ? styles.left : ''}`}>
+      <div onClick={this.onClick} className={`${styles.wrapper} ${className} ${wide ? styles.wide : ''} ${left ? styles.left : ''}`} {...rest}>
         <MaterialIcon className={styles.checkbox} icon={checked ? "check_box" : "check_box_outline_blank"} />
         { label && <label>{ label }</label> }
       </div>
